Miscellaneous documentation additions.
Added notes explaining how [expr, ..expr] form is used, with a pointer
to the `vec` and `at_vec` libraries for individuals like me who thought
it was more general and handled dynamic repeat expressions.

Added example of `do` syntax with a function of arity > 1; yes, one
should be able to derive this from the text above it, but it is still
a useful detail to compare and contrast against the arity == 1 case.

Added example of using for expression over a uint range, since someone
who is most used to write `for(int i; i < lim; i++) { ... }` will
likely want to know how to translate that form (regardless of whether
it happens to be good style or not for their use-case).

Added note about the semi-strange meaning of "fixed size" of vectors
in the vector type section.

I also have a couple of other additions in comments.  If someone is
reviewing these changes, I would appreciate feedback on whether I
should include those bits of text, or if they detract overall from the
document and thus should be removed.