Fixed various bugs in my port of lint.rs.  It seems to work in all cases now.
Including UnusedMutLintVisitor, which I rearranged the code of in an
attempt to improve its comparability with the original version in the
master branch.

The actual bug there, I think, was that visit_fn_action in
UnusedMutLintVisitor had a recursive traversal (the visit_*_action
methods are not supposed to traverse).  I suspect that there was
indeed a two-wrongs-make-a-right situation masking this bug earlier,
where the other wrong was an oversight (noted in next paragraph).

The TypeLimitsLintVisitor and part of UnusedMutLintVisitor was missing
the requiring OVERRIDE definitions in Visitor<..>; I probably should
document the pattern here.