add new enum ast::StrStyle as field to ast::lit_str
For the benefit of the pretty printer we want to keep track of how
string literals in the ast were originally represented in the source
code.

This commit changes parser functions so they don't extract strings from
the token stream without at least also returning what style of string
literal it was. This is stored in the resulting ast node for string
literals, obviously, for the package id in `extern mod = r"package id"`
view items, for the inline asm in `asm!()` invocations.

For `asm!()`'s other arguments or for `extern "Rust" fn()` items, I just
forbid raw string literals for now, mostly because it seemed
disproportionally cumbersome to thread that information through the
string processing that happens with those string literals, given the
limited advantage raw string literals would provide in these positions.

The other syntax extensions don't seem to store custom strings in the
ast, so they discard the style of strings they parse for now.