Revised shebang handling of comment gathering code (only used by pprust!).
This provides a better match in the pretty-printer for the semantics
of the lexer and parser: it only allows for a shebang as the very
first line it encounters (rather than checking for it on every line
during comment gathering).

Makes use of the newly added `lexer::is_at_shebang`, which introduced
the requirement that shebang comments follow the regexp `#!\s*[/\\]`
(that is, pound-bang-whitespace-slash).  The trailing slash reflects
the fact that (portable) shebangs need to use absolute paths.