Removed hacks on struct/enum/tuple.  Updated tests to reflect new reality.
In *practice* we will need to enforce that at most the last field in a
tuple/struct (or any variant of an enum) is unsized; it does not make
sense for anything preceding that to be unsized.  But I do not need to
fix that now to make forward progress elsewhere.

Updated the tests to reflect their intent (namely when the enum or
struct was supposed to be unsized), and to attempt to reflect *my*
intent (that eventually we will force things so that only the final
field is allowed to be unsized).