Adapted rt dump_registers function from graydon's gc branch. just playing around while reading graydon's old work on gc.
Expose Registers struct so that one can walk the registers in a
conservative GC scheme.

(but try to discourage forging of hand-made register sets by adding a
private NonCopyable member).

An important change to fix dump_registers was establishing the fxsave
area, as noted in the comment.

Also changed callback signature to take third argument (of type `*()`)
in an attempt to ensure that it will always be safe to tail-call from
a call to `dump_registers` over to the callback.

Adding tydesc argument to `exchange_malloc` to enable dynamic tracking
of ~T allocations, which in turn supports conservative scanning of
roots+stack holding pointers to ~T's and into their interiors.

Also, if we have `global_raw`, it is most consistent IMO to also
provide `free_raw`.

Checkpoint attempts to get things building atop
libgreen and revised TyDesc handling.