Attempting to hack in calls into BDW allocator from libstd and elsewhere.
My first go-round tried to use the non-debug variants of the calls,
but that yielded errors from host free routine saying that it was
being passed a ptr that had not been malloced.

So now I have switched to the debug variants, and am debugging the
problems that they are detecting (currently smashed objects warnings,
which I take as a sign that stuff is writing outside its bounds, but
that is not completely clear yet).