Added eager-buffer-expansion heuristic to Writer::write for MemWriter.
I put this in originally on my BDW branch to address some problems I
was seeing with excessive numbers of realloc calls with very poor
handling of the internal buffer.

(Better still could be to make a better prediction of the amount of
space needed when we allocate the buffer itself.  But I can wait
and/or live with something worse.)