Fix sample exprs for compatibility with current easy_syntax strategy.
In particular:
* All expressions must now be closed (no free variables),
* All expressions must have type unit (`()`).

Also, discovered issue 12643: labelled while loops are unsupported, so
I had to rewrite that example in terms of `loop` instead.  Made narrower
test case as I investigate this case.  At this point I am still failing
to construct control flow graphs for expressions with labelled break,
but this is a lot further than I was before.  I think.