Fixed problem I was seeing with control-flow jumps.
Fixed problem I was seeing with control-flow jumps appearing in odd
places in control-flow: the problem was that the analysis was treating
break and continue as if their predecessor jumped to the exit label,
while the break/continue expression itself fell through its successor.
This made the control-flow graphs hard-to-read.

(I also thought it made them incoherent, but on further reflection it
may have been an internally-consistent presentation of the flow;
e.g. the break expression itself *is* a jump, and so the things
following it sequentially are dead code, which is reflected in the
flow graph by making them dominated by the break expression itself.)

((Note that it may still have actually have been incoherent
nonetheless.  I thought I had worked out an example where the flow was
actively misleading.  Need to investigate further.))

I fixed the problem by passing along an additional bit of state along
with the CFGIndex returned by the recursive subroutines: the
combination is called a `Flow`, where a `Normal` flow may fall through
to its successor node, while a `Jumped` flow will never fall through.

I still need to double-check with Niko about whether he is comfortable
with this change to the architecture.