Added implementation of `loan_assert_matches`.
As part of this, introduces common abstract API unifying rustc's
dataflow and my local copy of dataflow.

As part of this, I gave up on trying to actually use my own local
version of the (semi-trivial) trait `DataFlowOperator`; it is much
easier to just reuse the abstraction provided by `rustc`.