Excuse regex from doing rustdoc tests atop stage1.
Hopefully final baby-step to getting `make check-stage1` working
again.

The use of a special-case check for regex here is ugly but is
analogous other similar checks for regex such as the one that landed
in PR #13844.

The way this is written, the user will get a reminder that
doc-crate-regex is being skipped whenever their rules attempt to do
the crate documentation tests.  This is deliberate: I want people
running `make check-stage1` to be reminded about which cases are being
skipped.  (But if such echo noise is considered offensive, it can
obviously be removed.)