Bug fixes to inverse_lookup routine to actually handle non-toplevel modules properly.
Added some instrumentation.

Also, the strategy used here is arugably pretty dumb, in that it will
linearly search the ast_map table multiple times to find each path
component, starting from the root, adding up to O(n*p) time for a path
with p parts and a crate of size n.

If we started by hoping that the leaf name is unique, we might end up
getting away with only taking O(n + p) time instead, by looking up the
leaf first and then just using the ast map to check that the parents
all match as you walk up from there.

But really both of these are terrible, so the most important thing for
now is to have code that is easy to understand.  I will ask people for
advise on how they would do this lookup tomorrow.