Add NodeId to `ast::Arm`.
The motivation for this is to delimit the scope of bindings introduced
by patterns on a match arm to just the expression and guard associated
with that arm.

It is possible that we may want to not to it this way, since I have a
vague impression that the rules are going to change so that when a
value moves into an arm's expression, the arm's guard will bind
by-reference, which may mean it will make more sense to just have
distinct representations for the two kinds of bindings that are
treated independenty by rustc's analyses.

But this was the easiest way to get it done in the short term.