rust-needsdrop: *proper* fix for long-standing bug with proc handling.
When doing `LoanPath::to_type`, `LpUpvar` is either an implicit
reference (when the upvar is captured by reference) or a value (when
captured by value and thus moved into the closure).

So I propagated this information into the `LoanPath` so that I can easily
compute the right thing in `LoanPath::to_type`.