FIXME rust-needsdrop checkpoint: mashup of foundational support for `trait QuietEarlyDrop`.
Most critical, generalized `lookup_vtable` to return a Result so that
we can use it as the basis for a `is_type_impl_of_trait` predicate.

There is a lot of other hackery mixed in here though (e.g. artifacts
of other false starts), so this commit needs some heavy history edit
love to trim it down to its cricial core.