Refactored librustc_trans to avoid stack-crossing ref cycle.
Namely, `BlockS` used to carry an `fcx: &'blk FunctionContext`, while
the `FunctionContext` carries a reference to the arena that holds the
blocks.

This creates a chicken/egg problem where we are attempting to assign
the lifetime `'blk` to both the `FunctionContext` and to the arena's
blocks, which does not work under the new lifetime rules for `Drop`,
since there has to be some order in which the two are torn down.

To resolve this, I removed the `fcx` from the `BlockS`, and instead
turn the `Block` type (which was formerly `&'blk BlockS`) into a struct
carrying both the pointer to the `BlockS` as well as the `fcx`.