REFACTORME Massive revision to trans::_match.
Added drop-flag updating when matching an input by move.

As part of this work, I spent a while making a `MatchInput` abstraction
that is a impoverished version of `Datum<Lvalue>`, and then threading
it through everywhere.

But now reviewing the outcome of this work, the main change being
introduced here (which may be far far far too conservative) did not
need the `MatchInput` change; it was entirely local to one function.