Change Lvalue ctors to take `bcx` and `node_id` for drop flags.
Made several of them start constructing non-trivial DropFlagInfo
values; namely `Lvalue::upvar`, `Lvalue::match_input`, and
`store_arg`.  (That last one is probably the most important, since
that covers fn parameters, one of the most important cases to cover.)

(Long term the policy decisions inlined in the code here could
probably be factored back out to the various call-sites...)

Revised the handling of match-pattern bindings to pass along a boolean
flag indicating whether it aliases other state or not.  (There is room
for improvement here; there are certainly cases that are currently
tagged as aliasing, but are part of match patterns that consume all of
the corresponding Lvalue input, and thus can safely be considered to
*not* alias other state, at least with respect to drop-filling.)