Add a debug-only fill-value, DTOR_HINTED, for dtors skipped due to hinting.
Added a handle_drop method to DropFlagInfo to abstract away which kind
of filling is done.

DTOR_HINTED is debug-only because in principle, no one should ever
read from the state of values that have been dropped, and thus it is
unobservable what value is written there -- but we also have sanity
checks double-checking whether the drop-flag has been established
correctly in debug builds. Therefore, we fill with DTOR_HINTED, and
then allow for that value in the sanity checks.

placate lints by removing now unused `must_zero` method.