refactor: distinguish between DataflowBuilder and Dataflow results.
In particular, distinguish between the DataflowBuilder (which, for
convenience, needs active access to the interpretation of the bits)
versus Dataflow results (which can just borrow such access for
routines like the graphviz pretty-printer).

This should allow me to reuse the same MoveData mutliple times as
input for distinct Dataflow analyses (which I plan to do in a
forthcoming commit).