add code to dataflow for establishing the on_entry for fn entry point.
this was not needed in the MovingOutStatements dataflow analysis, since
at entry, it suffices to say that no statements have reached that point.

But when the domain of the dataflow analysis is L-values, then we need
to establish up front that the MIR Arg L-values are initialized before
function entry.