Two new methods in `borrowck::mir::dataflow`.
`on_exit_set_for` is almost the dual to `on_entry_set_for` (in that it
lets you observe the bitvector on exit from the block rather than on
entry). Its interface is to return a `Vec` because the state is
constructed dynamically.

`fn each` is a hack of a method for folding the hopefully useful
operation of iteratively mapping each bit in a bitvector to its
interpretation and then visiting that interpretation. (Perhaps I
should pass along more state, like the index or something, but this is
good enough for my purposes for now.)