Add `pure_wrt_drop` field to `hir::TyParam` and `hir::LifetimeDef`.
The field is encoded via the `#[may_dangle]` attribute on the relevant
type or lifetime generic parameter.

Digression: I originally had the types carry a `ThinAttributes` field
instead, but the type `resolve_lifetime::ScopeChain` carries a
`hir::LifetimeDef` and is also used as the type for the static
variable `resolve_lifetime::ROOT_SCOPE`, so it cannot carry an
`Option<Vec<_>>`. So that forced me to map the attributes of interest
to lower level state in the `hir::LifetimeDef`. So it just seemed
consistent to do the same for `hir::TyParam`.)