FIXME: Allow `FunctionContext` in trans to carry around flow analysis results.
This is probably not the right thing. Redundantly re-running the flow
analysis extract the desired dataflow initialization data is probably
terrible.

TODO: Refactor so that the initialization data is computed and stored
somewhere once on the final MIR and then passed down to here
accordingly.

FIXME: Thread `FnKind` and `Span` through to `trans_fn` so it can borrowck.

something closer to compiling. factored flow results computation out from mir::borrowck.