Add an `Iterator<Item=T>` implementation for `IdxSet<T>`.
Note that this `iter` method is a little unusual as it needs to take
the index limit as an explicit argument. This is necessary because I
cannot encode this information in the data for `IdxSet<T>` itself.

(Maybe one day Rust will support user-defined dynamically-sized types
(DST's) that would let us attach a value like this to the metadata on
the DST reference.)