Switched `IdxSet::range` from a `Range<T>` param to a `Range<usize>` param.
The use of `Range<T>` for the `range` method on `IdxSet` was never
meaningful.  I.e. the main point of the range method is to allow one
to use a single bitvector to represent three sets held in sequence, so
it would be really weird to try to access arbitrary subranges of the
overall sequence.