An attempt to clarify the builder logic.
Before, we first matched on the type, and then only looked at the
projection when we needed to distinguish special cases. This might
have been reasonable, but I had a real hard time re-reading the code
and understanding it.

So now I instead match on both the projection-type and the type of the
base of the projection. (In most cases the two are intrinsically
linked.)