REMOVEME Attempted fix for missing EndRegion on borrowck-unary-move.rs.
While this might be good in some ways (it might reduce number of
unnecessary blocks being generated), it does not actually seem to fix
the problem at hand.

I'm keeping in in the commit queue for now, but should definitely
consider removing it once I have found the actual fix, and see whether
things are still fixed at that point.