REVISEME added `unwind` cache to Scope itself.
Removed previous hack to push EndRegions on top of already emitted blocks when "possible"

Didn't actually add code to make *use* of the newly added `unwind` cache.

Does not build yet because it asserts that the cache is unfilled whenever
you call `build_diverge_scope`, which apparently does not hold.