FIXME Make `--unpretty hir,identified` include both `NodeId` and `HirId`.
In particular, the numbers in things like the region scopes are
derived from the `HirId`, so those are more useful for understanding
some of the debug output.

The FIXME is a reminder that I should change this code to to print
only the `ItemLocalId`, not whole `HirId` (since in general the def-id
for the item just bloats the output wihtout adding much of use).