FIXME switch from `&fmt::Debug` trait object to pre-formmatted `DebugFormatted(String)`.
This resolves a long-standing annoyance that the debug-output helper
routine in `fn do_dataflow` was only able to return refrences into
(long-lived) structures, rather than having at least the option of
building up more interesting (if short-lived) debug output.

The FIXME is to remind myself that this alone won't actually build;
there are a couple changes to the commit that leaked from other work,
and some imports that are missing, etc.