BROKEN Checkpoint: Broken version of two-phase `&mut` dataflow.
The reason this is BROKEN is that the reservation and activation state
interact in a way that is not easily captured with the current
dataflow framework. (Namely, I want to only set the activation bit if
the corresponding reservation bit is already set. But this requires,
under the structure of this commit, the gen set for a block to be a
(potentially) non-trivial function of the entry-set state, rather than
a constant bitvector that we immediate bitwise-or with the entry-set
state.

My options are either: 1. try to generalize the dataflow framework to
support such non-trival gen-set construction, or 2. stop the madness
of the current approach and just split the analysis into two parts:
reservation and activation, with distinct bitvectors, where the
activation analysis will run after (and get to *read* the results
from) the reservation analysis.