Refactoring: avoid `tcx` explicitly through drop_flag_effects API.
Instead, it currently takes closures to compute the desired
value(s). I hope to replace the set of closures with two dedicated
traits to provide the methods has_uniform_drop_state and needs_drop,
but in a more self-documenting fashion.

In other words, this is exchanging passing around a trio of lifetimes
(which Rust currently cannot factor into any kind of abstract group)
for passing around one or more type parameters (which might, in
principle, be combined into a single type parameter, if one sets up
the traits the right way...)