Add `AutoBorrowMutability`; its like `hir::Mutability`  but w/ two-phase borrow info too.
Namely, the mutable borrows also carries a flag indicating whether
they should support two-phase borrows.

This will allow us to thread down, from the point of the borrow's
introduction, whether the particular adjustment that created it is one
that yields two-phase mutable borrows.

(Note that this particular commit is just a refactoring that inserts
the new enum and conversions between it and `hir::Mutability`; I have
not yet threaded it down to `rustc_mir::borrow_check`)