The qualify_consts transformation injects assignments for unassigned vars.
But it also causes an ICE when it encounters assignment statements in
const expressions.

I was previously injecting extra unassigned vars for a hypothetical
guard for all bindings: match arms without guards, and even function
arguments and let bindings. Those extra vars were then getting
assignments injected which subsequently caused an ICE.

This commit works around that by only injecting the extra vars for
match arms that have guards. (Arguably it would be even better to only
do it for guards in which the given variable occurs free, but I do not
feel like trying to implement that right now.)