FIXME experiment to only inject indirect ref on non `ref` match bindings.
I started on this path because I thought it might be simpler overall
than always trying to create an indirect ref to a ref (`& &T`) or
ref-mut (`& &mut T`).

However it, like many of my other attempts, still hits ill-formed MIR
panics, in this case because the MIR code generation is based on
manipulating a `&mut T` which we cannot readily get from the `&T` at
hand (in terms of type-checking).

So at this point I'm going back to thinking of doing indirect refs
everywhere and likewise derefs everywhere in the guard (so that the
temp for a `ref mut ident` binding has type `& &mut T`, but any
occurrence of ident are automatically derefed, yielding an expression
of type `&mut T` (and then the issue will presumably be caught by
MIR-borrowck).