FIXME thread the `errors_buffer` down into `nll::type_check`.
Right now its solely used for `check_local`, which ... I guess is not surprising?

This commit is tagged as "FIXME" because it explicitly imports the
`errors::Emit` trait since in one place it has a pretty dirty hack
where it dynamically detects whether we are buffering the errors or
not (based on whether it was passed a `Some(&mut errors_buffer)`).