FIXME First draft at better notes explaining the temp and sugggesting the semicolon.
fixme 1: niko would like us to say *why* we are making a temp (e.g. a
`&self` method call)

fixme 2: highlighting the source expr for the temp might be a bad idea
overall

fixme 3: might be good to try to avoid suggesting semicolon when the
types would not work out. (doing this completely may be harder than it
sounds; but perhps even simple heurisitcs would sufice.)