Dont duplicate entries in user annotations vector for every projection of a type.
This is a slight modification of PR 55937 and perhaps could/should be
cherry-picked to that PR.

Note that I am not doing this as a space optimization. I am doing it
for its semantic effect: having a single entry in the annotations
vector means that the different projections share that entry, and thus
share the inferred region variables for it. Prior to this change, each
projection would get its own set of inferred regions, which made it
hard to resolve issue 55748.