Turned on rustc side of report-future-incompat.
Also added support for testing it in `compiletest` (alternative was to just
discard the output, but I think we should check this in our test suite locally).

Note that the functionality is only enabled for the SAFE_PACKED_BORROWS lint due
to he allow-list in the recording function; that's why the set of altered tests
is so small.