Work around opaque types hiding needs_drop
When building the MIR we sometimes try to unschedule drops. In this we
assert that the drop has already been scheduled. Opaque types however
may be initialized with an expression kind that we know doesn't have a
type that needs to be dropped. To fix this we don't panic if we can't
find the drop of a variable with an opaque type.